requirements.txt
azure-identity == 1.15.0
azure-mgmt-storage == 21.1.0
azure-storage-blob == 12.8.0